document-preprocessor/requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
pytesseract
Pillow
fpdf
python-multipart
